docs: deepEqual使用方法补充
